fixed mapkicks after dcs (#BA2296) and bugging through walls by killing the aion process. also reduced leaveWorld delay to a minimum, so in case a connection loss gets detected after >= 10s player leaves world instantly. https://beyond-aion.com/forum/thread/2296-random-disconnects/

+++ game-server/src/com/aionemu/gameserver/services/player/PlayerLeaveWorldService.java
@@ -64,7 +64,7 @@ public class PlayerLeaveWorldService {
 	 * 
 	 * @see #leaveWorld(Player)
 	 */
-	public static void leaveWorldDelayed(Player player, int delayInMillis) {
+	public static void leaveWorldDelayed(Player player, long delayInMillis) {
 		Future<?> leaveWorldTask = ThreadPoolManager.getInstance().schedule(() -> leaveWorld(player), delayInMillis);
 		player.getController().addTask(TaskId.DESPAWN, leaveWorldTask);
 	}
